requirements.txt
pandas
matplotlib
numpy
tk
websocket-client
pynput
PyQt5
pyqtgraph
PyOpenGL 
PyOpenGL_accelerate
filterpy
pyserial
ahrs

 